Extract existing voting metadata unit tests
