+++ state_test.go
@@ -38,76 +38,6 @@ func createMessage(sender string, messageType MsgType, round ...uint64) *Message
 	return msg
 }
 
-func TestNodesCountVotingMetadata_MaxFaultyNodesCount(t *testing.T) {
-	cases := []struct {
-		TotalNodesCount, FaultyNodesCount uint
-	}{
-		{0, 0},
-		{1, 0},
-		{2, 0},
-		{3, 0},
-		{4, 1},
-		{5, 1},
-		{6, 1},
-		{7, 2},
-		{8, 2},
-		{9, 2},
-		{10, 3},
-		{99, 32},
-		{100, 33},
-	}
-	for _, c := range cases {
-		metadata := &NonWeightedVotingMetadata{nodesCount: c.TotalNodesCount}
-		assert.Equal(t, c.FaultyNodesCount, uint(metadata.MaxFaultyNodes()))
-	}
-}
-
-func TestNonWeightedVotingMetadata_QuorumSize(t *testing.T) {
-	cases := []struct {
-		TotalNodesCount, QuorumSize uint64
-	}{
-		{1, 1},
-		{2, 1},
-		{3, 1},
-		{4, 3},
-		{5, 3},
-		{6, 3},
-		{7, 5},
-		{8, 5},
-		{9, 5},
-		{10, 7},
-		{100, 67},
-	}
-
-	for _, c := range cases {
-		metadata := &NonWeightedVotingMetadata{nodesCount: uint(c.TotalNodesCount)}
-		assert.Equal(t, c.QuorumSize, metadata.QuorumSize())
-	}
-}
-
-func TestNonWeightedVotingMetadata_ValidNodesCount(t *testing.T) {
-
-	cases := []struct {
-		TotalNodesCount, RequiredMsgsCount int
-	}{
-		{1, 0},
-		{2, 0},
-		{3, 0},
-		{4, 2},
-		{5, 2},
-		{6, 2},
-		{7, 4},
-		{8, 4},
-		{9, 4},
-		{10, 6},
-		{100, 66},
-	}
-	for _, c := range cases {
-		metadata := &NonWeightedVotingMetadata{nodesCount: uint(c.TotalNodesCount)}
-		assert.Equal(t, c.RequiredMsgsCount, metadata.getRequiredMessagesCount())
-	}
-}
-
 func TestState_AddMessages(t *testing.T) {
 	pool := newTesterAccountPool()
 	validatorIds := []string{"A", "B", "C", "D"}
@@ -173,7 +103,7 @@ func TestState_MaxRound_Found(t *testing.T) {
 		}
 	}
 
-	maxRound, found := s.maxRound(&NonWeightedVotingMetadata{nodesCount: uint(s.validators.Len())})
+	maxRound, found := s.maxRound(NewVotingMetadata(nil, uint(s.validators.Len())))
 	assert.Equal(t, uint64(5), maxRound)
 	assert.Equal(t, true, found)
 }
@@ -191,7 +121,7 @@ func TestState_MaxRound_NotFound(t *testing.T) {
 	// Send wrong message type from some validator, whereas roundMessages map is empty
 	s.addMessage(createMessage(validatorIds[0], MessageReq_Preprepare))
 
-	metadata := &NonWeightedVotingMetadata{nodesCount: uint(s.validators.Len())}
+	metadata := NewVotingMetadata(nil, uint(s.validators.Len()))
 	maxRound, found := s.maxRound(metadata)
 	assert.Equal(t, maxRound, uint64(0))
 	assert.Equal(t, found, false)
